refactor: moved createStore & storeList
