+++ client/app/components/group/_createStore/_createStoreMap/createStoreMap.html
