+++ client/app/components/group/_createStore/createStore.js
@@ -3,8 +3,8 @@ import uiRouter from "angular-ui-router";
 import ngMaterial from "angular-material";
 import createStoreComponent from "./createStore.component";
 import createStoreMap from "./_createStoreMap/createStoreMap";
-import Store from "../../common/store/store";
-import Geocoding from "../../common/geocoding/geocoding";
+import Store from "../../services/store/store";
+import Geocoding from "../../services/geocoding/geocoding";
 
 let createStoreModule = angular.module("createStore", [
   uiRouter,
