+++ client/app/components/group/_storeList/_storeMap/storeMap.controller.js
