+++ client/app/components/group/_storeList/_storeMap/storeMap.html
