+++ client/app/components/group/_storeList/_storeMap/storeMap.js
