+++ client/app/components/group/_storeList/_storeMap/storeMap.spec.js
