+++ client/app/components/group/_storeList/storeList.component.js
