+++ client/app/components/group/_storeList/storeList.controller.js
