+++ client/app/components/group/_storeList/storeList.spec.js
