+++ client/app/components/group/_storeList/storeList.styl
